bug fixes on hyperparameter space of other classifiers and feature preprocessors

+++ autosklearn/pipeline/components/classification/libsvm_svc.py
@@ -90,7 +90,7 @@ class LibSVM_SVC(AutoSklearnClassificationAlgorithm):
         else:
             self.coef0 = float(self.coef0)
         self.tol = float(self.tol)
-        self.max_iter = float(self.max_iter)
+        self.max_iter = int(self.max_iter)
 
         self.shrinking = check_for_bool(self.shrinking)
